Only compare TYPE bits in Scheduler
Not using the ARCH bits to search for compatible FPGA task accelerators as the new task created may have more than 1 archtecture. Also, submit the tasks to spawnOut when the FPGA arch bit is not set, instead of doing so when the SMP bit is set. Improve the module code avoiding magic numbers